+++ CHANGES.md
@@ -70,6 +70,7 @@
 -   pat checklist: Also dispatch standard ``change`` event when de/selecting all items.
 -   pat select: Add missing ``<span>`` element around the select element itself. Fixes: https://github.com/quaive/ploneintranet.prototype/issues/1087
 -   pat depends/core utils: Do not set inline styles when showing elements in transition mode ``none``. Fixes #719.
+-   pat scroll: Fix scrolling offset incorrectly applied. Fixes: #763.
 
 
 ## 3.0.0-dev - unreleased